+++ testsuite/integration/manualmode/src/test/java/org/jboss/as/test/manualmode/web/valve/HelloServlet.java
@@ -31,16 +31,19 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.jboss.logging.Logger;
 
 @WebServlet(name="HelloServlet", urlPatterns={"/"})
 public class HelloServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
+    private static Logger log = Logger.getLogger(HelloServlet.class);
 
     public void doGet(HttpServletRequest request,
                       HttpServletResponse response)
         throws IOException, ServletException {
         PrintWriter out = response.getWriter();
         out.println("HelloServlet");
+        log.info("HelloServlet");
     }
 
 }
