+++ testsuite/integration/manualmode/src/test/java/org/jboss/as/test/manualmode/web/valve/MANIFEST.MF
@@ -1,2 +1,2 @@
 Manifest-Version: 1.0
-Dependencies: org.jboss.testvalve
+Dependencies: org.jboss.testvalve.standard
