+++ testsuite/integration/manualmode/src/test/java/org/jboss/as/test/manualmode/web/valve/authenticator/web-global-auth.xml
@@ -17,7 +17,7 @@
 -->    
     <servlet>
         <servlet-name>HelloServlet</servlet-name>
-        <servlet-class>org.jboss.as.test.manualmode.web.valve.authenticator.HelloServlet</servlet-class>
+        <servlet-class>org.jboss.as.test.manualmode.web.valve.HelloServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
